apply refactor changes
