+++ packages/svelte-docgen/src/kind/mod.js
@@ -1,0 +1,13 @@
+/**
+ * JavaScript namespace `is` declaration for type guards.
+ * @module
+ */
+
+import { isTypeRef, isAdvancedType, isBaseType, isInstantiableType } from "./guard.js";
+
+export {
+	isTypeRef as typeRef,
+	isAdvancedType as advancedType,
+	isBaseType as baseType,
+	isInstantiableType as instantiableType,
+};
