+++ packages/svelte-docgen/src/parser/type/type-parameter.test.ts
@@ -3,7 +3,7 @@ import { describe, it } from "vitest";
 import { create_options } from "../../../tests/shared.js";
 import type * as Doc from "../../doc/type.js";
 import { parse } from "../mod.js";
-import { isTypeRef } from "../../doc/utils.js";
+import { isTypeRef } from "../../kind/guard.js";
 
 describe("TypeParam", () => {
 	const { props, types } = parse(
